Introduce a DOME module and an exit function
There is now a DOME module which can contain DOME-centric configuration
methods and properties. It also adds an `exit` method which can be used
by a game to exit the execution.

A caveat to this is that you can't quit a VM from within the C code,
so we need to wait for the execution to reach the core engine loop
before we can safely shutdown the Wren VM.