[#5] The DOME.exit method won't wait til end of frame
We use setjmp/longjmp to provide a way for the WrenVM to jump to the end
of the game loop in order to cease execution.

You can't exit the WrenVM from within the VM call stack, and even
aborting the fibre isn't guarenteed to finish. setjmp let's us do that
by bypassing the rest of the VM execution.