AudioEngine is now a static class
The original implementation of AudioEngine set it up as a foreign class
in order to take advantage of Wren's allocate/finalize paradigm, but
this meant we had to do various things to give the impression that the
AudioEngine was in fact static.

Instead, we can just have the main DOME engine manage the memory and
eliminate those shims all together.

This means we don't worry about users accidentally trying to use
multiple audio engines or other things, as well as removing the
unnecessary cruft of an 'init' module.