fix #2: DOME can play unlimited audio channels
We had a fixed limit of 4-8 audio channels which could be played at
once, because I didn't know how to resize memory properly.

Now we can resize the channel list on the fly and thus play all queued
channels at once. This makes it the game's responsibility to filter out
audio which shouldn't be played.